+++ pipelined/bevy_render2/src/render_graph/graph.rs
@@ -313,7 +313,7 @@ mod tests {
     }
 
     #[test]
-    pub fn test_graph_edges() {
+    fn test_graph_edges() {
         let mut graph = RenderGraph::default();
         let a_id = graph.add_node("A", TestNode::new(0, 1));
         let b_id = graph.add_node("B", TestNode::new(0, 1));
@@ -369,7 +369,7 @@ mod tests {
     }
 
     #[test]
-    pub fn test_get_node_typed() {
+    fn test_get_node_typed() {
         struct MyNode {
             value: usize,
         }
@@ -401,7 +401,7 @@ mod tests {
     }
 
     #[test]
-    pub fn test_slot_already_occupied() {
+    fn test_slot_already_occupied() {
         let mut graph = RenderGraph::default();
 
         graph.add_node("A", TestNode::new(0, 1));
@@ -421,7 +421,7 @@ mod tests {
     }
 
     #[test]
-    pub fn test_edge_already_exists() {
+    fn test_edge_already_exists() {
         let mut graph = RenderGraph::default();
 
         graph.add_node("A", TestNode::new(0, 1));