+++ pipelined/bevy_sprite2/src/rect.rs
@@ -1,4 +1,3 @@
-use bevy_core::Byteable;
 use bevy_math::Vec2;
 
 /// A rectangle defined by two points. There is no defined origin, so 0,0 could be anywhere
@@ -21,5 +20,3 @@ impl Rect {
         self.max.y - self.min.y
     }
 }
-
-unsafe impl Byteable for Rect {}